- feature: added locatorTimeout - chore: moved mocha stuff out od steps dir
