+++ packages/testing/src/steps/index.ts
@@ -1,5 +1,3 @@
 export * from './steps';
 export * from './predicates';
 export * from './types';
-export * from './mocha-ctx';
-export * from './time-dilation';
